refactor(frontend): component to display media search item
